namiz/requirements.txt
eve
boto3
python-dotenv
loguru
dynaconf
pymongo
pytest
moto
mongomock
pytest-cov
